[Behat][API] Extract some steps to separate contexts to resolve duplications
