+++ src/Sylius/Behat/Resources/config/suites/api/promotion/applying_catalog_promotions.yml
@@ -25,11 +25,12 @@ default:
                 - Sylius\Behat\Context\Transform\CatalogPromotionContext
 
                 - sylius.behat.context.api.admin.managing_catalog_promotions
-                - sylius.behat.context.api.admin.managing_product_variants
                 - sylius.behat.context.api.admin.save
                 - sylius.behat.context.api.shop.login
                 - sylius.behat.context.api.shop.product
                 - sylius.behat.context.api.shop.product_variant
+                - Sylius\Behat\Context\Api\Admin\CreatingProductVariantContext
+                - Sylius\Behat\Context\Api\Admin\ManagingProductVariantsPricesContext
                 - Sylius\Behat\Context\Api\Admin\ManagingProductTaxonsContext
 
             filters:
